Extended msess storage and manager software.
Further work with the msess storage. Added functions for removing
measurement sessions and probes. Wrote code for flushing measurement
data do SQLite-database, so far untested. Also begun integrating
the msess-storage with the rest of the code.

Improved the config-module of the manager, now it can successfully
be instanciated from any other class with the same state allover,
similar to the logger-module. Also finished the getParam-function
to return config parameter from an xpath-string.