Begun working towards IPC.
Began the work towards the "new" architecture. Less action in the C
program, more in the python part. Removed almost all probe handling
in msess as it will be handled by the python part.

Modified the protocol handler to iterate the msess-list and send
packets to all members when needed.

Began adding functionality to the config handler to populate the msess
storage list.

WARNING: DOES NOT COMPILE! Quick n' dirty commit as I'm on my way...