Measurement session configuration from config file.
Probed now reads the measurement sessions from XML and populates the
msess storage from it. The msess list is then iterated to create
client forks. The msess list is the used to create client forks
as well as keeping track of what destinations to send probes to as well
as the probe transmission interval.