Data aggregation and error handling.
Continued the work with data aggregation and remote procedures.
There is now a fully working API to fetch measurement data over
XML-RPC, but its painfully slow. Needs to be modified heavily.

Also improved error handling. The thread stopping is now run in
try-except clauses to catch errors such as variables not set yet
(if for example a constructor throws an exception before the
other thread objects have been created).

The probed-process is also beein monitored and restarted if it
disappears.