Fetch config from the web, more args to probed.
The configuration file is now fetched from the web and saved
to disk. More options to probed are fetched from the config
file (such as port and timestamping mode).

There was also added a function to get storage statistics from
the ProbeStore via XML-RPC to simplify debugging of high
memory usage.

Also modified the function which deletes old data from the
database to hopefully work...