Keep track of state for highres-data.
Keeps track of each measurement session's state (up/down) to avoid
filling the ASM with high-resolution data when a node is down.

When the number of packets of state pongloss or timeout is equal
to the total number of packets, the node is considered as being
down. High-resolution data will be saved when a session's state
has changed or when the state is up, but we have lost packets.