Added errchecker.py - simplifies performance tests
errchecker.py connects to SLA-NG nodes and fetches measurement data from
them. It is then examined to determine it the measurements are failing
or not - that is we have packet loss or not. It also prints out some
general measurement statistics in a easily overviewed format.

What devices and what session ID:s to check is configured in a dict in
the file itself.