Updated diff_ts function to handle negative values
The diff_ts function handles negative values in a better way; the value
written to the return ts_t is the absolute value. The sign is given by
the return value of the function. It also returns correct results for
all valid input values, which the old implementation did not.

It is probably quite inefficient though, I'll save that for a rainy day
:)

Fixes #3